+++ src/services/RTCService.ts
@@ -1,27 +1,44 @@
 import logger from '../util/logger'
+import * as Storage from '../util/storage'
+import { isDefined } from '../util/helpers'
 import { DialogOptions, ICacheDevices, ICacheResolution } from '../interfaces/'
 
-const getUserMedia = async (constraints: MediaStreamConstraints) => {
+const getUserMedia = async (constraints: MediaStreamConstraints): Promise<MediaStream | null> => {
   logger.debug('RTCService.getUserMedia', constraints)
+  const { audio, video } = constraints
+  if (!audio && !video) {
+    return null
+  }
   const stream = await navigator.mediaDevices.getUserMedia(constraints).catch(error => error)
   if (streamIsValid(stream)) {
     return stream
   }
-  logger.error('RTCService.getUserMedia', stream)
+  logger.error('getUserMedia error: ', stream.name, stream.message)
   throw stream
 }
 
-const streamIsValid = stream => stream instanceof MediaStream
+const streamIsValid = (stream: MediaStream) => stream && stream instanceof MediaStream
 
-// TODO: cache the devices in localStorage if available!
 const getDevices = async (): Promise<ICacheDevices> => {
   const devices = await navigator.mediaDevices.enumerateDevices()
-    .catch(error => { logger.error('enumerateDevices Error', error) })
-  const cache = {}
+    .catch(error => {
+      logger.error('enumerateDevices Error', error)
+      return null
+    })
+  const cache = {};
+  ['videoinput', 'audioinput', 'audiooutput'].map((kind: string) => {
+    cache[kind] = {}
+    Object.defineProperty(cache[kind], 'toArray', {
+      value: function () {
+        return Object.keys(this).map(k => this[k])
+      }
+    })
+  })
   if (devices) {
-    devices.forEach(t => {
+    devices.forEach((t: MediaDeviceInfo) => {
       if (!cache.hasOwnProperty(t.kind)) {
-        cache[t.kind] = {}
+        logger.warn(`Unknown device type: ${t.kind}`, t)
+        return true
       }
       if (t.groupId && Object.keys(cache[t.kind]).some(k => cache[t.kind][k].groupId == t.groupId)) {
         return true
@@ -29,6 +46,9 @@ const getDevices = async (): Promise<ICacheDevices> => {
       cache[t.kind][t.deviceId] = t
     })
   }
+
+  Storage.setItem('devices', cache)
+
   return cache
 }
 
@@ -40,7 +60,7 @@ const getResolutions = async (): Promise<ICacheResolution[]> => {
     const constraints = { audio: false, video: { width: { exact: resolution[0] }, height: { exact: resolution[1] } } }
     const stream = await getUserMedia(constraints).catch(error => null)
     if (stream) {
-      stream.getVideoTracks().forEach(t => {
+      stream.getVideoTracks().forEach((t: MediaStreamTrack) => {
         supported.push(Object.assign({ resolution: `${resolution[0]}x${resolution[1]}` }, t.getSettings()))
         t.stop()
       })
@@ -55,7 +75,7 @@ const getMediaConstraints = (options: DialogOptions): MediaStreamConstraints => 
     if (typeof audio === 'boolean') {
       audio = {}
     }
-    audio['deviceId'] = { exact: options.micId }
+    audio.deviceId = { exact: options.micId }
   }
 
   let { video = false } = options
@@ -63,10 +83,65 @@ const getMediaConstraints = (options: DialogOptions): MediaStreamConstraints => 
     if (typeof video === 'boolean') {
       video = {}
     }
-    video['deviceId'] = { exact: options.camId }
+    video.deviceId = { exact: options.camId }
   }
 
   return { audio, video }
+}
+
+const assureDeviceId = async (id: string, label: string, kind: MediaDeviceInfo['kind']): Promise<string> => {
+  const devices = await navigator.mediaDevices.enumerateDevices()
+    .catch(error => { logger.error('enumerateDevices Error', error) })
+  if (devices) {
+    for (let i = 0; i < devices.length; i++) {
+      const { deviceId, label: deviceLabel, kind: deviceKind } = devices[i]
+      if (kind !== deviceKind || !/input$/.test(kind)) {
+        continue
+      }
+      if ((id && id === deviceId) || (label && label === deviceLabel)) {
+        return deviceId
+      }
+    }
+  }
+  return null
+}
+
+const checkPermissions = async (): Promise<boolean> => {
+  const fullStream = await getUserMedia({ audio: true, video: true }).catch(_e => null)
+  if (streamIsValid(fullStream)) {
+    fullStream.getTracks().forEach((t: MediaStreamTrack) => t.stop())
+  } else {
+    const audioStream = await getUserMedia({ audio: true }).catch(_e => null)
+    if (streamIsValid(audioStream)) {
+      audioStream.getTracks().forEach((t: MediaStreamTrack) => t.stop())
+    } else {
+      return false
+    }
+  }
+  return true
+}
+
+const removeUnsupportedConstraints = (constraints: MediaTrackConstraints): void => {
+  const supported = navigator.mediaDevices.getSupportedConstraints()
+  Object.keys(constraints).map(key => {
+    if (!supported.hasOwnProperty(key)) {
+      logger.warn(`"${key}" constraint is not supported in this browser!`)
+      delete constraints[key]
+    }
+  })
+}
+
+const checkDeviceIdConstraints = async (id: string, label: string, kind: MediaDeviceInfo['kind'], constraints: MediaTrackConstraints) => {
+  const { deviceId } = constraints
+  if (!isDefined(deviceId) && (id || label)) {
+    const deviceId = await assureDeviceId(id, label, kind).catch(error => null)
+    if (deviceId) {
+      constraints.deviceId = { exact: deviceId }
+    } else {
+      throw `Unknown device with id: '${id}' and label: '${label}'`
+    }
+  }
+  return constraints
 }
 
 export {
@@ -74,5 +149,9 @@ export {
   getDevices,
   getResolutions,
   getMediaConstraints,
-  streamIsValid
+  streamIsValid,
+  assureDeviceId,
+  checkPermissions,
+  removeUnsupportedConstraints,
+  checkDeviceIdConstraints
 }